+++ lib/simple_bayes/storage/behaviour.ex
@@ -1,4 +1,5 @@
 defmodule SimpleBayes.Storage.Behaviour do
   @callback init(%SimpleBayes{}, Keyword.t) :: pid
-  @callback load(Keyword.t) :: pid
+  @callback save(pid, Keyword.t)            :: {:ok, pid}
+  @callback load(Keyword.t)                 :: pid
 end